+++ test/config/with_float_convert.rb
@@ -1,0 +1,1 @@
+max_fast_inline Float::INFINITY