+++ History.md
@@ -2,6 +2,7 @@
 
 * Features
   * Your feature goes here <Most recent on the top, like GitHub> (#Github Number)
+  *  Adds max_fast_inline as a configuration option for the Server object (#2406)
 
 * Bugfixes
   * Cleanup daemonization in rc.d script (#2409)